Add $service_up values (#69)
* cleanup: rearrange imports

* tests: setup mocks/teardown outside suite

The current code does all the setup of mocks and teardown in the suite
level which prevents us from doing any negative testing (the state of
the mocks is preserved through the entire suite instead of test).

* add service_up value

This allows the prometheus endpoint to expose if the past scrape had a
failure or it was successful.  This can help in checking if the specific
service has any current on-going issues where it can't be reached.

* refactor: move error logging up a layer

Given that we pass the error up now to the top-most level, we can
actually log the error inside the layer above the exporter and simply
pass the error up.

* prom: make the test *parse* values

* cleanup: drop redundant log messages

in exporter.go, we already publish a note of what metrics we're
about to collect so it's a tad bit useless to log it twice.

* refactor: drop extra layer

We're calling back up to exporter.CollectMetrics with nothing more, so
we can just drop it and call exporter.Collect() which is directly
defined in the main object.

* refactor: move describe to parent

Describe() does the same thing everwhere, so let's move it back up.

* refactor: use prometheus.Collector

Prometheus provides it's own interface for a Collector so let's
use that.

* cleanup: rearrange imports

* tests: setup mocks/teardown outside suite

The current code does all the setup of mocks and teardown in the suite
level which prevents us from doing any negative testing (the state of
the mocks is preserved through the entire suite instead of test).

* add service_up value

This allows the prometheus endpoint to expose if the past scrape had a
failure or it was successful.  This can help in checking if the specific
service has any current on-going issues where it can't be reached.

* refactor: move error logging up a layer

Given that we pass the error up now to the top-most level, we can
actually log the error inside the layer above the exporter and simply
pass the error up.

* prom: make the test *parse* values

* cleanup: drop redundant log messages

in exporter.go, we already publish a note of what metrics we're
about to collect so it's a tad bit useless to log it twice.

* refactor: drop extra layer

We're calling back up to exporter.CollectMetrics with nothing more, so
we can just drop it and call exporter.Collect() which is directly
defined in the main object.

* refactor: move describe to parent

Describe() does the same thing everwhere, so let's move it back up.

* refactor: use prometheus.Collector

Prometheus provides it's own interface for a Collector so let's
use that.

* drop extra println