Simplify unit testing and test values
This patch greatly simplifies the unit testing by using tools
provided by the testutils package from Prometheus and a simple
way to define expected values.

It also adds all the values based on the current mocks, implements
a few missing ones and makes the test suite fail if it hits an
endpoint that is not mocked.

It also took care of a few limit and missing mocks which were
not being properly and fully tested.