Add ip_version to IP availablities
When trying to alert and check if you have IPs available, you need
to be able to group by the IP version because if you have both an
IPv4 and IPv6 subnet, you could be running out of IPv4 addresess
in a network but have plenty of IPv6 left.

This patch adds it and updates the appropriate tests.