Add openstack_neutron_port
Instead of presenting aggregates of data, we should expose all the data
that we have and let our users instead using things like recording rules
to aggregate them.

This patch also refactors the Neutron bits to not make multiple requests
to all ports which puts a huge strain on the cluster.