glance: apiUP to monitor api availiblity
the import-method API call will help to make sure the glance
API is alive and healthy without get images list.
If we have hundreds of images and want to get a list from
the glance API it takes too long and it doesn't worth to
monitor the count of images, we just need to make sure glance API is alive.

Signed-off-by: Parsa <p.yousefi97@gmail.com>