Move snap readme to snap directory
Note that snapcraft complains about it,
but it doesn't appear to cause issues:

```
The 'snap' directory is meant specifically for snapcraft, but it contains
the following non-snapcraft-related paths:
- README.md

This is unsupported and may cause unexpected behavior. If you must store
these files within the 'snap' directory, move them to 'snap/local'
which is ignored by snapcraft.
```