Use CPU info instead of hypervisor list data
The total number of VCPUs should be obtained from CPUInfo instead
of the VCPUs column of hypervisor list. This value changes if there
are pinned, dedicated or shared CPUs configured in nova.

The CPUInfo data may be empty in some openstack versions, so
fallback to old method in that case.